Makefile and Linkdef.h update

+++ Analysis/TreeFilter/Core/python/core.py
@@ -117,25 +117,28 @@ def ParseArgs() :
     parser.add_argument('--disableOutputTree', dest='disableOutputTree', default=False, action='store_true', help='Do not write events to the output tree')
 
     parser.add_argument('--copyInputFiles', dest='copyInputFiles', default=False, action='store_true', help='Transfer files to local directory. Only works with batch.')
-    
+
     #-----------------------------
     # Occasionally used
     #-----------------------------
 
     parser.add_argument('--sample', dest='sample', default=None, help='Name of sample.  May be used in cases where the sample name must be known to the c++ code')
-    
+
     parser.add_argument('--moduleArgs', dest='moduleArgs', default=None, help='Arguments to pass to module.  Should be in the form of a dictionary')
 
     parser.add_argument('--write_file_list', dest='write_file_list', default=False, action='store_true', help='Write a text file containing files found on eos.  use --read_file_list to use this file in the future')
 
     parser.add_argument('--read_file_list', dest='read_file_list', default=False, action='store_true', help='Read the file list instaed of looking on eos.  Must have used --write_file_list to create the file' )
-    
+
     return parser.parse_args()
 
 
 
 
 
+#-------------------------------------------
+#          Main config function
+#-------------------------------------------
 
 def config_and_run( options, package_name ) :
 
@@ -292,25 +295,30 @@ def config_and_run( options, package_name ) :
 
         os.mkdir( '/tmp/%s' %tmpname )
 
-        # Write the c++ files having the branch definitions and 
+        # Write the c++ files having the branch definitions and
         # SetBranchAddress calls
-        # Write all output branches in the 
+        # Write all output branches in the
         # header file so that the code will compile
         # only the keep branches will be saved, however
-        #write_header_files(new_brdef_file_name, new_linkdef_file_name, branches,out_branches, [ br['name'] for br in branches ], alg_list )
-        write_header_files(new_brdef_file_name, new_linkdef_file_name, branches,out_branches, branches_to_keep, alg_list )
+        write_header_files(new_brdef_file_name , branches,out_branches, branches_to_keep, alg_list )
 
         write_source_file(new_source_file_name, new_header_file_name, branches, out_branches, branches_to_keep, write_expert_code=options.writeExpertCode )
 
         if not (filecmp.cmp(  old_brdef_file_name, new_brdef_file_name ) and
                 filecmp.cmp(  old_header_file_name, new_header_file_name ) and
-                filecmp.cmp( old_source_file_name, new_source_file_name )  and
-                filecmp.cmp( old_linkdef_file_name, new_linkdef_file_name ) )  :
+                filecmp.cmp( old_source_file_name, new_source_file_name ) ) :
 
             print '--------------------------------------'
             print 'Difference found in processing code'
             print 'Will use new code and will compile'
             print '--------------------------------------'
+            if not filecmp.cmp( old_brdef_file_name,  new_brdef_file_name ) :
+                os.system("diff %s %s" % ( old_brdef_file_name,  new_brdef_file_name  ))
+            if not filecmp.cmp( old_header_file_name, new_header_file_name ) :
+                os.system("diff %s %s" % ( old_header_file_name, new_header_file_name ))
+            if not filecmp.cmp( old_source_file_name, new_source_file_name ) :
+                os.system("diff %s %s" % ( old_source_file_name, new_source_file_name ))
+
 
             # If we need to recompile then a different executable should
             # be made so as to not overwrite an existing one
@@ -607,8 +615,9 @@ def compile_code( alg_list, branches, out_branches, branches_to_keep, workarea, 
     # Write all output branches in the 
     # header file so that the code will compile
     # only the keep branches will be saved, however
-    #write_header_files(brdef_file_name, linkdef_file_name, branches,out_branches, [ br['name'] for br in branches ], alg_list )
-    write_header_files(brdef_file_name, linkdef_file_name, branches,out_branches, branches_to_keep, alg_list )
+    #write_header_files(brdef_file_name, branches,out_branches, [ br['name'] for br in branches ], alg_list )
+    write_header_files(brdef_file_name , branches, out_branches, branches_to_keep, alg_list )
+    write_linkdef_file(linkdef_file_name, branches )
 
     write_source_file(source_file_name, header_file_name, branches, out_branches, branches_to_keep, write_expert_code=writeExpertCode )
 
@@ -731,7 +740,7 @@ def get_branch_map_from_tree( tree ) :
         title = title.replace(']]', '][')
  
         lf = br.GetListOfLeaves()[0]
-        type  = lf.GetTypeName()
+        typen  = lf.GetTypeName()
         size  = lf.GetLen()
         is_range = lf.IsRange()
         # match any number of brackets and get them.  
@@ -794,7 +803,7 @@ def get_branch_map_from_tree( tree ) :
         if varId is not None :
             leafEntry += '/%s' %varId
 
-        prop_dic = {'name' : name, 'type' : type, 'totSize' : totSize, 'arrayStr' : arrayStr, 'leafEntry' : leafEntry, 'sizeEntries' : size_entries }
+        prop_dic = {'name' : name, 'type' : typen, 'totSize' : totSize, 'arrayStr' : arrayStr, 'leafEntry' : leafEntry, 'sizeEntries' : size_entries }
         if name in [d['name'] for d in all_branches] :
             continue
 
@@ -1164,11 +1173,11 @@ def get_file_evt_map( input_files, nsplit, nFilesPerJob, totalEvents, treeName ,
     # then each file runs over the full set of events
     files_evtsplit = []
 
-    # determine how many times to split each file.  Do this by 
+    # determine how many times to split each file.  Do this by
     # adding a split to each file until the number
     # of splits is achieved
     files_nsplit = [ 1 ]*len(split_files)
-    n_addtl_split = nsplit - len(split_files) 
+    n_addtl_split = nsplit - len(split_files)
     files_idx = 0
     for sp in range(0, n_addtl_split ) :
         files_nsplit[files_idx] += 1
@@ -1510,7 +1519,7 @@ def create_job_desc_file(command_info, kwargs) :
 #
 #    return desc_file_name
 
-def write_header_files( brdefname, linkdefname, branches, out_branches=[], keep_branches=[], alg_list=[] ) :
+def write_header_files( brdefname, branches, out_branches=[], keep_branches=[], alg_list=[] ) :
 
     branch_header = open(brdefname, 'w')
     branch_header.write('#ifndef BRANCHDEFS_H\n')
@@ -1577,6 +1586,7 @@ def write_header_files( brdefname, linkdefname, branches, out_branches=[], keep_
 
     branch_header.close()
 
+def write_linkdef_file( linkdefname, branches ) :
     # this file is used to create root dictionaries
     # for vector<vector<>> types.  It will always be
     # created and if no such types exist it will
@@ -1586,26 +1596,61 @@ def write_header_files( brdefname, linkdefname, branches, out_branches=[], keep_
     link_types = []
     for conf in branches :
         if conf['type'].count('vector') > 0 :
-            modtype = conf['type'].replace('vector','std::vector')
+            modtype = conf['type'].replace('vector','std::vector')\
+                                  .replace(' ','')
             link_types.append( modtype )
 
     # remove duplicates
     link_types = set( link_types )
 
-    link_header = open(linkdefname, 'w')
-    link_header.write('#ifdef __CINT__\n')
-    link_header.write('#pragma link off all globals;  \n')
-    link_header.write('#pragma link off all classes;  \n')
-    link_header.write('#pragma link off all functions;\n')
-    link_header.write('#pragma link C++ nestedclasses;\n\n')
+    if os.path.isfile(linkdefname):
+        link_header_orig = open(linkdefname, 'r')
+        original_lines = link_header_orig.readlines()
+        original_lines_reduced = "".join(original_lines).replace(" ",'')
+        link_header_orig.close()
 
-    for type in link_types :
-        link_header.write('#pragma link C++ class %s+;\n' %type)
+        ## find out which link types are not already in the old file
+        new_link_types = [ typ for typ in link_types if "class"+typ not in original_lines_reduced ]
 
-    link_header.write('#endif\n\n')
+        if new_link_types:
+            print "new types found: ", " ".join(new_link_types)
+            print "will recompile Linkdef.h"
+            ## find out where to insert new lines
+            link_header = open(linkdefname, 'w')
+            for i, line in enumerate(original_lines):
+                if "endif" in line:
+                    ## insert new pragmas before endif
+                    for typ in new_link_types :
+                        link_header.write('#pragma link C++ class %s+;\n' %typ)
+                link_header.write(line)
+            link_header.close()
+            proc = subprocess.Popen(['make', 'linkdef'])
+            proc.wait()
+        else:
+            print "no new types"
+    else:
+        ## if Linkdef.h does not exist, make a new one
+        print "Linkdef.h does not exist"
+        print "link types to be included: ", " ".join(new_link_types)
 
-    link_header.close()
-    
+        link_header = open(linkdefname, 'w')
+        link_header.write('#ifdef __CINT__\n')
+        link_header.write('#pragma link off all globals;\n')
+        link_header.write('#pragma link off all classes;\n')
+        link_header.write('#pragma link off all functions;\n')
+        link_header.write('#pragma link C++ nestedclasses;\n\n')
+
+        for typ in link_types :
+            link_header.write('#pragma link C++ class %s+;\n' %typ)
+
+        link_header.write('#endif\n')
+
+        link_header.close()
+
+        # generate dictionary file
+        proc = subprocess.Popen(['make', 'linkdef'])
+        proc.wait()
+
 
 def write_source_file(source_file_name, header_file_name, branches, out_branches=[], keep_branches=[], debugCode=False, write_expert_code=False) :
